migration complete with tools and zero shot
