+++ backend/app/app/api/v1/endpoints/search.py
@@ -1,0 +1,9 @@
+from fastapi import APIRouter
+
+from app.utils.typesense_func import client
+
+router = APIRouter()
+
+@router.get("")
+def search(q:str):
+    return q
