+++ backend/app/app/utils/chains.py
@@ -1,5 +1,5 @@
 from langchain import LLMChain, PromptTemplate
-from langchain.chat_models import ChatOpenAI
+from langchain_community.chat_models import ChatOpenAI
 
 import re
