+++ backend/app/app/utils/tools.py
@@ -1,14 +1,29 @@
 from app.core.config import settings
+from app.models.search import TypesenseSearchResult
+from app.utils.typesense_func import client
 from langchain.chat_models import ChatOpenAI
 from langchain.schema import HumanMessage
-from langchain.tools import BaseTool
+from langchain.tools import BaseTool, tool
 from serpapi import GoogleSearch
 from typing import Any, Optional
 import httpx
+import os
 
-pokemon_api_url = "https://pokeapi.co/api/v2/pokemon/"
 unsplash_api_url = f"https://api.unsplash.com/search/photos?client_id={settings.UNSPLASH_API_KEY}&query="
-weather_api_url = "https://wttr.in"
+
+@tool
+def anime_search_tool(input: str) -> dict[str, Any]:
+    """use this tools when you need to lookup information about anime. parameter are input. input should the name of the keyword to search"""
+    WEBPAGE_INDEX = os.environ.get("WEBPAGE_INDEX")
+    search_parameters = {
+        "q": input,
+        "query_by": "title",
+        "per_page": 2,
+    }
+    search_hits = client.collections[WEBPAGE_INDEX].documents.search(search_parameters)
+    entity_result = TypesenseSearchResult(**search_hits)
+    hits = entity_result.hits
+    return hits[0].document if hits else []
 
 
 class GeneralKnowledgeTool(BaseTool):
@@ -28,54 +43,6 @@ class GeneralKnowledgeTool(BaseTool):
         response = await chat.agenerate([[HumanMessage(content=query)]])
         message = response.generations[0][0].text
         return message
-
-
-class PokemonSearchTool(BaseTool):
-    name = "search_pokemon"
-    description = " Useful when asked to answer information about a pokemon"
-
-    def __init__(self):
-        super().__init__()
-
-    def _run(self, query: str, run_manager: Optional[Any] = None) -> str:
-        """Use the tool."""
-        response = httpx.get(pokemon_api_url + query)
-        body = response.json()
-        pokemon_number = body["id"]
-        pokemon_name = body["name"]
-        return f"{pokemon_name} - #{pokemon_number} "
-
-    async def _arun(self, query: str, run_manager: Optional[Any] = None) -> str:
-        """Use the tool asynchronously."""
-        async with httpx.AsyncClient() as client:
-            pokemon_url = pokemon_api_url + query.lower()
-            response = await client.get(pokemon_url)
-            if response.status_code == 404:
-                return "Pokemon not found"
-            body = response.json()
-            pokemon_number = body["id"]
-            pokemon_name = body["name"]
-            pokemon_image = body.get("sprites", {}).get(
-                "front_default",
-                "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/25.png",
-            )
-            pokemon_types = body["types"]
-            pokemon_type = pokemon_types[0].get("type", {}).get("name", "electric")
-            abilities = body.get("abilities", [])
-            abilities_list_names = []
-            height = body["height"] / 10  # convert to meters
-            weight = body["weight"] / 10  # convert to kg
-            for ability in abilities:
-                abilities_list_names.append(ability["ability"]["name"])
-            abilities_names = ", ".join(abilities_list_names)
-            return f"""{pokemon_image} \n
-{pokemon_name} - #{pokemon_number} \n
-Abilities: {abilities_names} \n
-Type: {pokemon_type} \n
-Height: {height} m \n
-Weight: {weight} kg \n
-Sprite: {pokemon_image} \n
-"""
 
 
 class ImageSearchTool(BaseTool):
@@ -143,29 +110,3 @@ class YoutubeSearchTool(BaseTool):
                 ]
             )
             return video_list_string
-
-
-class GeneralWeatherTool(BaseTool):
-    name = "Weather"
-    description = "useful for when you need to answer questions about weather"
-
-    def __init__(self):
-        super().__init__()
-        # self.return_direct = True
-
-    def _run(self, query: str, run_manager: Optional[Any] = None) -> str:
-        """Use the tool."""
-        pass
-
-    async def _arun(self, query: str, run_manager: Optional[Any] = None) -> dict:
-        """Use the tool asynchronously."""
-        async with httpx.AsyncClient() as client:
-            query = query.replace(" ", "+")
-            query = query.replace(",", "")
-            final_url = f"{weather_api_url}/{query}?format=j1"
-            response = await client.get(final_url)
-            if response.status_code == 404:
-                return f"Could not find weather for {query}"
-            weather = response.json()
-            temperature = weather["current_condition"][0]
-            return temperature
