+++ backend/app/app/utils/typesense_func.py
