+++ backend/app/pyproject.toml
@@ -11,7 +11,7 @@ include = "app"
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
 mypy = "^1.5.0"
-ruff = "^0.0.284"
+ruff = "^0.2.1"
 black = "^23.7.0"
 
 [tool.black]
@@ -41,10 +41,12 @@ fastapi = {version = "^0.99.1", extras = ["all"]}
 fastapi-pagination = {version = "^0.12.7", extras = ["sqlalchemy"]}
 asyncer = "^0.0.2"
 httpx = "^0.24.1"
-langchain = "^0.0.265"
+langchain = "^0.1.7"
 openai = "^0.27.8"
 adaptive-cards-py = "^0.0.7"
 google-search-results = "^2.4.2"
+ruff = "^0.2.1"
+langchainhub = "^0.1.14"
 
 [tool.ruff.per-file-ignores]
 "__init__.py" = [ "F401",]
