--- fe/sample-result.py
@@ -1,5 +1,0 @@
-{'document': {'id': 'People/99356', 'image_url': 'https://s4.anilist.co/file/anilistcdn/staff/medium/4356.jpg', 'name': 'Maki Naruto', 'type': ['People'], 'url': 'https://anilist.co/staff/99356'},
-'highlight': {'name': {'matched_tokens': ['Naruto'], 'snippet': 'Maki <mark>Naruto</mark>'}},
-'highlights': [{'field': 'name', 'matched_tokens': ['Naruto'], 'snippet': 'Maki <mark>Naruto</mark>'}],
-'text_match': 578730123365187705,
-'text_match_info': {'best_field_score': '1108091338752', 'best_field_weight': 15, 'fields_matched': 1, 'score': '578730123365187705', 'tokens_matched': 1}}
