--- pyproject.toml
@@ -1,18 +1,0 @@
-[tool.poetry]
-name = "zenzen-ask"
-version = "0.1.0"
-description = ""
-authors = ["Lim Jian Huei <jhueilim96@gmail.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-openai2 = "^1.6.7"
-contextcrunch-langchain = "^1.0.2"
-fastapi-restify = "^0.0.4"
-streamlit1 = "^0.1.0"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
