--- requirements.in
@@ -1,179 +1,0 @@
-adaptive-cards-py==0.0.7
-aiohttp==3.9.3
-aiosignal==1.3.1
-altair==5.2.0
-annotated-types==0.6.0
-anyio==3.7.1
-appnope==0.1.4
-asttokens==2.4.1
-async-timeout==4.0.3
-asyncer==0.0.2
-attrs==23.2.0
-black==23.12.1
-blinker==1.7.0
-blis==0.7.11
-build==1.0.3
-cachecontrol==0.13.1
-cachetools==5.3.2
-catalogue==2.0.10
-certifi==2024.2.2
-cffi==1.16.0
-charset-normalizer==3.3.2
-cleo==2.1.0
-click==8.1.7
-cloudpathlib==0.16.0
-colorama==0.4.6
-comm==0.2.1
-confection==0.1.4
-crashtest==0.4.1
-create-fastapi-project==0.2.4
-cymem==2.0.8
-dataclasses-json==0.5.14
-debugpy==1.8.1
-decorator==5.1.1
-distlib==0.3.8
-distro==1.9.0
-dnspython==2.5.0
-dulwich==0.21.7
-email-validator==2.1.0.post1
-exceptiongroup==1.2.0
-executing==2.0.1
-fastapi==0.99.1
-fastapi-pagination==0.12.15
-fastjsonschema==2.19.1
-filelock==3.13.1
-frozenlist==1.4.1
-gitdb==4.0.11
-gitpython==3.1.42
-google-search-results==2.4.2
-h11==0.14.0
-httpcore==0.17.3
-httptools==0.6.1
-httpx==0.24.1
-idna==3.6
-importlib-metadata==7.0.1
-iniconfig==2.0.0
-installer==0.7.0
-itsdangerous==2.1.2
-jaraco-classes==3.3.1
-jedi==0.19.1
-jinja2==3.1.3
-jsonpatch==1.33
-jsonpointer==2.4
-jsonschema==4.21.1
-jsonschema-specifications==2023.12.1
-jupyter-client==8.6.0
-jupyter-core==5.7.1
-keyring==24.3.0
-langchain==0.0.265
-langchain-community==0.0.20
-langchain-core==0.1.23
-langchainhub==0.1.14
-langcodes==3.3.0
-langsmith==0.0.87
-markdown-it-py==3.0.0
-markupsafe==2.1.5
-marshmallow==3.20.2
-matplotlib-inline==0.1.6
-mdurl==0.1.2
-more-itertools==10.2.0
-msgpack==1.0.7
-multidict==6.0.5
-murmurhash==1.0.10
-mypy==1.8.0
-mypy-extensions==1.0.0
-nest-asyncio==1.6.0
-numexpr==2.9.0
-numpy==1.26.4
-openai==0.27.10
-openapi-schema-pydantic==1.2.4
-orjson==3.9.14
-packaging==23.2
-pandas==2.2.0
-parso==0.8.3
-pathspec==0.12.1
-pexpect==4.9.0
-pillow==10.2.0
-pkginfo==1.9.6
-platformdirs==3.11.0
-pluggy==1.4.0
-poetry==1.7.1
-poetry-core==1.8.1
-poetry-plugin-export==1.6.0
-preshed==3.0.9
-prompt-toolkit==3.0.36
-protobuf==4.25.3
-psutil==5.9.8
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pyarrow==15.0.0
-pycparser==2.21
-pydantic==1.10.14
-pydantic-core==2.14.6
-pydeck==0.8.0
-pygments==2.17.2
-pyproject-hooks==1.0.0
-pytest==7.4.4
-python-dateutil==2.8.2
-python-dotenv==1.0.1
-python-multipart==0.0.9
-pytz==2024.1
-pyyaml==6.0.1
-pyzmq==25.1.2
-questionary==2.0.1
-rapidfuzz==3.6.1
-referencing==0.33.0
-requests==2.31.0
-requests-toolbelt==1.0.0
-rich==13.7.0
-rpds-py==0.18.0
-ruff==0.0.284
-setuptools==69.1.0
-shellingham==1.5.4
-six==1.16.0
-smart-open==6.4.0
-smmap==5.0.1
-sniffio==1.3.0
-spacy==3.7.3
-spacy-legacy==3.0.12
-spacy-loggers==1.0.5
-sqlakeyset==2.0.1706067133
-sqlalchemy==2.0.27
-srsly==2.4.8
-stack-data==0.6.3
-starlette==0.27.0
-streamlit==1.31.1
-strenum==0.4.15
-tenacity==8.2.3
-thinc==8.2.2
-thinc-apple-ops==0.1.4
-toml==0.10.2
-tomli==2.0.1
-tomlkit==0.12.3
-toolz==0.12.1
-tornado==6.4
-tqdm==4.66.2
-traitlets==5.14.1
-trove-classifiers==2024.1.31
-typer==0.9.0
-types-requests==2.31.0.20240125
-typesense==0.19.0
-typing-extensions==4.9.0
-typing-inspect==0.9.0
-tzdata==2024.1
-tzlocal==5.2
-ujson==5.9.0
-urllib3==2.2.0
-uv==0.1.1
-uvicorn==0.27.1
-uvloop==0.19.0
-validators==0.22.0
-virtualenv==20.25.0
-wasabi==1.1.2
-watchfiles==0.21.0
-wcwidth==0.2.13
-weasel==0.3.4
-websockets==12.0
-xattr==0.10.1
-yarl==1.9.4
-zipp==3.17.0
